requirements.txt
flask~=3.0.3
flask-cors
requests~=2.31.0
python-dotenv~=1.0.1
team-14-lib-version